RAMDISK: sched: Set up Zygote properly for EAS
The stock HTC 10 Android N ramdisk only set up Zygote to
write PIDs to stune/foreground.

We add the same lines but for top-app to set it up for top-app.

Bug: this method somehow breaks writing PIDs to foreground.
It isn't a big deal for now as we aren't boosting foreground PIDs.
However, prefer idle won't be working, again nothing major.

This fixes Zygote properly writing PIDs for top-app/tasks in stune.
This is vital to have boosting working for top-app.